requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==6.1.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
asttokens==3.0.1
    # via stack-data
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==7.0.6
    # via streamlit
certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via streamlit
cryptography==46.0.7
    # via google-auth
db-dtypes==1.4.4
    # via urban-resilience-pipeline (pyproject.toml)
decorator==5.2.1
    # via ipython
executing==2.2.1
    # via stack-data
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-api-core==2.30.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.49.2
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-bigquery==3.41.0
    # via urban-resilience-pipeline (pyproject.toml)
google-cloud-core==2.5.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-storage==3.10.1
    # via urban-resilience-pipeline (pyproject.toml)
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.80.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.80.0
    # via google-api-core
idna==3.12
    # via requests
ipython==9.12.0
    # via urban-resilience-pipeline (pyproject.toml)
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
matplotlib-inline==0.2.1
    # via ipython
narwhals==2.20.0
    # via
    #   altair
    #   plotly
numpy==2.4.4
    # via
    #   db-dtypes
    #   pandas
    #   pydeck
    #   streamlit
packaging==26.1
    # via
    #   altair
    #   db-dtypes
    #   google-cloud-bigquery
    #   plotly
    #   streamlit
pandas==3.0.2
    # via
    #   urban-resilience-pipeline (pyproject.toml)
    #   db-dtypes
    #   streamlit
parso==0.8.6
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==12.2.0
    # via streamlit
plotly==6.7.0
    # via urban-resilience-pipeline (pyproject.toml)
prompt-toolkit==3.0.52
    # via ipython
proto-plus==1.27.2
    # via google-api-core
protobuf==6.33.6
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
py4j==0.10.9.9
    # via pyspark
pyarrow==24.0.0
    # via
    #   db-dtypes
    #   streamlit
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.13.3
    # via urban-resilience-pipeline (pyproject.toml)
pydantic-core==2.46.3
    # via pydantic
pydeck==0.9.2
    # via streamlit
pygments==2.20.0
    # via
    #   ipython
    #   ipython-pygments-lexers
pyspark==4.1.1
    # via urban-resilience-pipeline (pyproject.toml)
python-dateutil==2.9.0
    # via
    #   google-cloud-bigquery
    #   pandas
python-dotenv==1.2.2
    # via urban-resilience-pipeline (pyproject.toml)
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.1
    # via
    #   urban-resilience-pipeline (pyproject.toml)
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.3
    # via gitdb
stack-data==0.6.3
    # via ipython
streamlit==1.56.0
    # via urban-resilience-pipeline (pyproject.toml)
tenacity==9.1.4
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.5
    # via streamlit
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   altair
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.2.1
    # via requests
wcwidth==0.6.0
    # via prompt-toolkit
